# back-end/docs/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/revenue-ledger/summary/platform": {
      "get": {
        "operationId": "RevenueLedgerController_getPlatformSummary",
        "parameters": [
          {
            "name": "x-id",
            "in": "header",
            "description": "Caller user ID (UUID)",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "total / pending / disbursed platform amounts"
          }
        },
        "security": [
          {
            "x-role": []
          },
          {
            "bearer": []
          }
        ],
        "summary": "Platform-wide revenue summary — super_user only",
        "tags": [
          "Revenue Ledger"
        ]
      }
    },
    "/revenue-ledger/my": {
      "get": {
        "operationId": "RevenueLedgerController_getMyEarnings",
        "parameters": [
          {
            "name": "x-id",
            "in": "header",
            "description": "Caller user ID (UUID)",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          },
          {
            "bearer": []
          }
        ],
        "summary": "Get current user earnings",
        "tags": [
          "Revenue Ledger"
        ]
      }
    },
    "/revenue-ledger/provider/{spId}": {
      "get": {
        "operationId": "RevenueLedgerController_getProviderEarnings",
        "parameters": [
          {
            "name": "x-id",
            "in": "header",
            "description": "Caller user ID (UUID)",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "spId",
            "required": true,
            "in": "path",
            "description": "sp_id UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          },
          {
            "bearer": []
          }
        ],
        "summary": "Provider earnings summary — SP (own), RM, or super_user",
        "tags": [
          "Revenue Ledger"
        ]
      }
    },
    "/revenue-ledger/region-manager/{rmId}": {
      "get": {
        "operationId": "RevenueLedgerController_getRmEarnings",
        "parameters": [
          {
            "name": "x-id",
            "in": "header",
            "description": "Caller user ID (UUID)",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "rmId",
            "required": true,
            "in": "path",
            "description": "rm_id UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          },
          {
            "bearer": []
          }
        ],
        "summary": "Region Manager earnings — RM (own) or super_user",
        "tags": [
          "Revenue Ledger"
        ]
      }
    },
    "/revenue-ledger/{id}/payout": {
      "patch": {
        "operationId": "RevenueLedgerController_dispatchPayout",
        "parameters": [
          {
            "name": "x-id",
            "in": "header",
            "description": "Caller user ID (UUID)",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ledger_id UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          },
          {
            "bearer": []
          }
        ],
        "summary": "Dispatch payout for a PENDING ledger entry — super_user only",
        "tags": [
          "Revenue Ledger"
        ]
      }
    },
    "/revenue-ledger/{id}": {
      "get": {
        "operationId": "RevenueLedgerController_findOne",
        "parameters": [
          {
            "name": "x-id",
            "in": "header",
            "description": "Caller user ID (UUID)",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ledger_id UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          },
          {
            "bearer": []
          }
        ],
        "summary": "Get ledger entry by ID — super_user only",
        "tags": [
          "Revenue Ledger"
        ]
      },
      "delete": {
        "operationId": "RevenueLedgerController_remove",
        "parameters": [
          {
            "name": "x-id",
            "in": "header",
            "description": "Caller user ID (UUID)",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ledger_id UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          },
          {
            "bearer": []
          }
        ],
        "summary": "Hard delete a ledger entry — super_user only",
        "tags": [
          "Revenue Ledger"
        ]
      }
    },
    "/revenue-ledger": {
      "post": {
        "operationId": "RevenueLedgerController_create",
        "parameters": [
          {
            "name": "x-id",
            "in": "header",
            "description": "Caller user ID (UUID)",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRevenueLedgerDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          },
          {
            "bearer": []
          }
        ],
        "summary": "Manually create ledger entry — super_user only",
        "tags": [
          "Revenue Ledger"
        ]
      },
      "get": {
        "operationId": "RevenueLedgerController_findAll",
        "parameters": [
          {
            "name": "x-id",
            "in": "header",
            "description": "Caller user ID (UUID)",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          },
          {
            "bearer": []
          }
        ],
        "summary": "List all ledger entries (scoped for managers)",
        "tags": [
          "Revenue Ledger"
        ]
      }
    },
...
  },
  "info": {
    "title": "TatkuUnited API",
    "description": "Service platform API",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Paste your JWT token (without the Bearer prefix)"
      },
      "x-role": {
        "type": "apiKey",
        "in": "header",
        "name": "x-role",
        "description": "Role header required by RolesGuard. Allowed values: super_user, region_manager, service_provider, customer"
      }
    },
    "schemas": {
...
      "CreateRevenueLedgerDto": {
        "type": "object",
        "properties": {
          "booking_id": {
            "type": "string",
            "example": "bkg-123"
          },
          "service_id": {
            "type": "string",
            "example": "svc-123"
          },
          "sp_id": {
            "type": "string",
            "example": "sp-123"
          },
          "rm_id": {
            "type": "string",
            "example": "rm-123"
          },
          "provider_amount": {
            "type": "number",
            "example": 850
          },
          "platform_amount": {
            "type": "number",
            "example": 150
          },
          "payout_status": {
            "type": "string",
            "example": "PENDING",
            "default": "PENDING"
          }
        },
        "required": [
          "booking_id",
          "service_id",
          "sp_id",
          "rm_id",
          "provider_amount",
          "platform_amount"
        ]
      },
...
    }
  }
}